Pass command buffer list to WebGPU queue as argument

+++ src/types.zig
@@ -1,1 +1,1 @@
-pub const Id = usize;
+pub const Id = u32;
